Move lib test to tests directory
- It is effectively an integration test
